remove mean from Numeric and refactor file structure
