+++ sample_size/sample_size_calculator/variables.py
@@ -20,10 +20,8 @@ class Boolean:
 class Numeric:
     def __init__(
         self,
-        mean: float,
         variance: float,
     ):
-        self.mean = mean
         self.variance = variance
